Implemented Income Creation

+++ app/src/main/java/com/autonomyway/AutonomyWayFacade.java
@@ -5,7 +5,7 @@ import com.autonomyway.model.Income;
 import java.util.List;
 
 public interface AutonomyWayFacade {
-    Income createIncome(String name, int recurrentTime, int recurrentCash, Income.Type type);
+    Income createIncome(String name, long recurrentTime, long recurrentCash, Income.Type type);
 
     List<Income> getIncomeList();
